I need to check this in now.  It probably works right now.  I'll continue working on this at home this evening.

+++ platform/emulator/weakdict.cc
@@ -4,28 +4,121 @@
 #include "tagged.hh"
 #include "am.hh"
 
-static WeakDictionary* gcLinkedList = 0;
+// apparently we have to be really careful not to perform a deref
+// during gc
+inline int _oz_isWeakDictionary(OZ_Term t)
+{
+  return oz_isExtension(t) &&
+    oz_tagged2Extension(t)->getIdV()==OZ_E_WEAKDICTIONARY;
+}
+
+inline WeakDictionary* _tagged2WeakDictionary(OZ_Term t)
+{
+  Assert(_oz_isWeakDictionary(t));
+  return (WeakDictionary*) oz_tagged2Extension(t);
+}
+
+inline int oz_isWeakDictionary(OZ_Term t)
+{
+  return _oz_isWeakDictionary(OZ_deref(t));
+}
+
+inline WeakDictionary* tagged2WeakDictionary(OZ_Term t)
+{
+  return _tagged2WeakDictionary(OZ_deref(t));
+}
+
+inline bool WeakDictionary::isEmpty()
+{
+  return (table==0 || table->numelem==0);
+}
+
+// GARBAGE COLLECTION SUPPORT
+
+// this is a list of currently known weak dictionaries
+// it is computed anew at each GC by copying.  creating
+// a new weak dictionary also adds it to this list.
+
+static OZ_Term weakList = 0;
+
+// at the very start of GC, we must save the list of currently
+// known weak dictionaries.  We can then process this list at
+// any point prior to finalization provided we also check for
+// GC marks.
+
+static OZ_Term previousWeakList;
+
+void gCollectWeakDictionariesInit()
+{
+  previousWeakList = weakList;
+  weakList = 0;
+}
+
+// All previously known dictionaries that
+//      (1) have a live board
+//      (2) have a stream
+//      (3) are not empty
+// must be preserved because they must still be given a chance to
+// eventually finalize their content.
+
+void gCollectWeakDictionariesPreserve()
+{
+  // we simply go through the list of previously known
+  // dictionaries and collect those that have not yet been
+  // collected and that verify the 3 conditions above
+
+  if (previousWeakList==0) return;
+  for (;previousWeakList!=AtomNil;
+       previousWeakList=oz_tail(previousWeakList))
+    {
+      OZ_Term t = oz_head(previousWeakList);
+      WeakDictionary*d = _tagged2WeakDictionary(t);
+      if (! (((int32)d->__getSpaceInternal())&1) && /* not already gc marked */
+          ((Board*)d->__getSpaceInternal())->cacIsAlive() &&
+          d->stream!=0 &&
+          !d->isEmpty())
+        // gc has the side effect of entering the copy into
+        // weakList, thus recording its existence for the
+        // next time
+        oz_gCollectTerm(t,t);
+    }
+}
+
+void gCollectWeakDictionariesContent()
+{
+  // This is called after the 1st gc phase has completed.
+  // all data reachable outside of weak dictionaries has
+  // been marked and copied.  Entries in a weak dictionary
+  // now point either to marked values (that have already
+  // been copied to the new heap) or to unmarked values
+  // (which are no longer reachable from outside).
+  if (weakList==0) return;
+  for (OZ_Term curr=weakList;curr!=AtomNil;curr=oz_tail(curr))
+    _tagged2WeakDictionary(oz_head(curr))->weakGC();
+}
+
+OZ_Extension* WeakDictionary::gCollectV()
+{
+  // copy to the new heap and enter into known list
+  // WeakDictionary's gcRecurseV does nothing.
+  // the real worked is done in gCollectWeakDictionaries.
+  WeakDictionary* d = new WeakDictionary(table,stream);
+  if (weakList==0) weakList=AtomNil;
+  weakList = oz_cons(OZ_extension(d),weakList);
+  return d;
+}
+
+OZ_Extension* WeakDictionary::sCloneV() {
+  return gCollectV();
+}
+
+// EXTENSION INTERFACE
 
 OZ_Term WeakDictionary::printV(int depth)
 {
   return oz_pair2(oz_atom("<WeakDictionary n="),
                   oz_pair2(oz_int(table->numelem),
                            oz_atom(">")));
-}
-
-OZ_Extension* WeakDictionary::gCollectV()
-{
-  // copy to the new heap and enter into linked list
-  // WeakDictionary's gcRecurseV does nothing.
-  // the real worked is done in gCollectWeakDictionaries.
-  WeakDictionary* d = new WeakDictionary(table,stream);
-  d->next = gcLinkedList;
-  gcLinkedList = d;
-  return d;
-}
-
-OZ_Extension* WeakDictionary::sCloneV() {
-  return NULL;
 }
 
 inline OZ_Boolean WeakDictionary::get(OZ_Term key,OZ_Term& val)
@@ -62,45 +155,19 @@ OZ_Return WeakDictionary::putFeatureV(OZ_Term f,OZ_Term  v)
   return PROCEED;
 }
 
-void gCollectWeakDictionaries()
-{
-  // This is called after the 1st gc phase has completed.
-  // all data reachable outside of weak dictionaries has
-  // been marked and copied.  Entries in a weak dictionary
-  // now point either to marked values (that have already
-  // been copied to the new heap) or to unmarked values
-  // (which are no longer reachable from outside).  The
-  // weak dictionaries to be processed have been linked
-  // together in gcLinkedList.
-  for (;gcLinkedList;gcLinkedList=gcLinkedList->next)
-    gcLinkedList->weakGC();
-  // now gcLinkedList==0 again
-}
+// BUILTINS
 
 OZ_BI_define(weakdict_new,0,2)
 {
-  if (!OZ_onToplevel())
-    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
-  OZ_Term srm = oz_newFuture(oz_rootBoard());
+  OZ_Term srm = oz_newFuture(oz_currentBoard());
   WeakDictionary* wd = new WeakDictionary(srm);
   OZ_out(0) = srm;
   OZ_out(1) = OZ_extension(wd);
+  if (weakList==0) weakList=AtomNil;
+  weakList = oz_cons(OZ_out(1),weakList);
   return PROCEED;
 }
 OZ_BI_end
-
-inline int oz_isWeakDictionary(OZ_Term t)
-{
-  t = OZ_deref(t);
-  return OZ_isExtension(t) &&
-    OZ_getExtension(t)->getIdV()==OZ_E_WEAKDICTIONARY;
-}
-
-inline WeakDictionary* tagged2WeakDictionary(OZ_Term t)
-{
-  Assert(oz_isWeakDictionary(t));
-  return (WeakDictionary*) OZ_getExtension(OZ_deref(t));
-}
 
 #define OZ_declareWeakDict(ARG,VAR) \
 OZ_declareType(ARG,VAR,WeakDictionary*,"weakDictionary",\
@@ -113,9 +180,9 @@ OZ_declareType(ARG,VAR,OZ_Term,"feature",OZ_isFeature,NO_COERCE)
 
 OZ_BI_define(weakdict_put,3,0)
 {
- if (!OZ_onToplevel())
-    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
  OZ_declareWeakDict(0,d);
+ if (!d->isLocal())
+    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
  OZ_declareDetTerm(1,k);
  OZ_declareTerm(2,v);
  TaggedRef w = v;
@@ -135,8 +202,6 @@ OZ_BI_define(weakdict_get,2,1)
 {
   OZ_declareWeakDict(0,d);
   OZ_declareFeature(1,k);
-  if (!OZ_onToplevel() && !d->isLocal())
-    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
   OZ_Term v;
   if (!d->get(k,v))
     return oz_raise(E_SYSTEM,E_KERNEL,"weakDictionary",2,OZ_in(0),OZ_in(1));
@@ -148,8 +213,6 @@ OZ_BI_define(weakdict_condGet,3,1)
 {
   OZ_declareWeakDict(0,d);
   OZ_declareFeature(1,k);
-  if (!OZ_onToplevel() && !d->isLocal())
-    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
   OZ_Term v;
   if (!d->get(k,v)) OZ_RETURN(OZ_in(2));
   OZ_RETURN(v);
@@ -175,7 +238,7 @@ void WeakDictionary::close()
 OZ_BI_define(weakdict_close,0,0)
 {
   OZ_declareWeakDict(0,d);
-  if (!OZ_onToplevel() && !d->isLocal())
+  if (!d->isLocal())
     return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
   d->close();
   return PROCEED;
@@ -218,11 +281,6 @@ OZ_BI_define(weakdict_items,1,1)
 }
 OZ_BI_end
 
-inline bool WeakDictionary::isEmpty()
-{
-  return (table==0 || table->numelem==0);
-}
-
 OZ_BI_define(weakdict_isempty,1,1)
 {
   OZ_declareWeakDict(0,d);
@@ -260,6 +318,8 @@ inline void WeakDictionary::remove(OZ_Term key)
 OZ_BI_define(weakdict_remove,2,0)
 {
   OZ_declareWeakDict(0,d);
+  if (!d->isLocal())
+    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
   OZ_declareFeature(1,key);
   d->remove(key);
   return PROCEED;
@@ -277,6 +337,8 @@ inline void WeakDictionary::remove_all()
 OZ_BI_define(weakdict_remove_all,1,0)
 {
   OZ_declareWeakDict(0,d);
+  if (!d->isLocal())
+    return oz_raise(E_ERROR,E_KERNEL,"globalState",1,oz_atom("weakDictionary"));
   d->remove_all();
   return PROCEED;
 }
